+++ changelog.md
@@ -1,2 +1,2 @@
 ### Benched 1.2.2
-- Initial release of Benched for Minecraft 1.21.9 & 1.21.10
+- Initial release of Benched for Minecraft 1.21.11