+++ changelog.md
@@ -1,2 +1,4 @@
+## Update to Minecraft 26.1
+
 ### Benched 1.2.2
 - Initial release of Benched for Minecraft 1.21.11